pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==23.2.1",
    "fastapi==0.104.1",
    "httpx==0.25.2",
    "itsdangerous>=2.2.0",
    "jinja2==3.1.2",
    "pyrogram==2.0.106",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "tgcrypto==1.2.5",
    "uvicorn[standard]==0.24.0",
]
